refactor: split agent runtime state
